fix(web): small ui-changes
